+++ apex/contrib/bottleneck/__init__.py
@@ -1,1 +1,2 @@
 from .bottleneck import Bottleneck, SpatialBottleneck
+from .halo_exchangers import HaloExchangerNoComm, HaloExchangerAllGather, HaloExchangerSendRecv
